requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0
chardet==5.2.0
python-dateutil==2.8.2
pydantic==2.5.0
openpyxl==3.1.2
numpy==1.25.2
requests==2.31.0
